Finish class. alg module. Bugfix in pipeline test code
A bunch of scikit learn classification algorithms are now wrapped and
good to go in the algorithm_classification.py module, and there are
wrappers for simple hyperparameter optimization (really simple) and for
registering the algs in a global registry at definition. This allows you
to add algorithms by just adding code in one single place.